Refactor to move duplicated methods/constant into module
