+++ app/vacancy_sources/vacancy_source/source/broadbean.rb
@@ -1,22 +1,9 @@
 class VacancySource::Source::Broadbean
   include VacancySource::Parser
+  include VacancySourceShared
 
   FEED_URL = ENV.fetch("VACANCY_SOURCE_BROADBEAN_FEED_URL").freeze
   SOURCE_NAME = "broadbean".freeze
-  EXCLUDED_DETAILED_SCHOOL_TYPES = [
-    "Further education",
-    "Other independent school",
-    "Online provider",
-    "British schools overseas",
-    "Institution funded by other government department",
-    "Miscellaneous",
-    "Offshore schools",
-    "Service children’s education",
-    "Special post 16 institution",
-    "Other independent special school",
-    "Higher education institutions",
-    "Welsh establishment",
-  ].freeze
 
   # Helper class for less verbose handling of items in the feed
   class FeedItem
@@ -134,10 +121,6 @@ class VacancySource::Source::Broadbean
       readable_job_location: first_school&.name,
       about_school: first_school&.description,
     }
-  end
-
-  def vacancy_listed_at_excluded_school_type?(schools)
-    (schools.map(&:detailed_school_type) & EXCLUDED_DETAILED_SCHOOL_TYPES).present?
   end
 
   def find_schools(item)
