+++ app/vacancy_sources/vacancy_source/source/fusion.rb
@@ -1,22 +1,9 @@
 class VacancySource::Source::Fusion
   include VacancySource::Parser
+  include VacancySourceShared
 
   FEED_URL = ENV.fetch("VACANCY_SOURCE_FUSION_FEED_URL").freeze
   SOURCE_NAME = "fusion".freeze
-  EXCLUDED_DETAILED_SCHOOL_TYPES = [
-    "Further education",
-    "Other independent school",
-    "Online provider",
-    "British schools overseas",
-    "Institution funded by other government department",
-    "Miscellaneous",
-    "Offshore schools",
-    "Service children’s education",
-    "Special post 16 institution",
-    "Other independent special school",
-    "Higher education institutions",
-    "Welsh establishment",
-  ].freeze
 
   class FusionImportError < StandardError; end
 
@@ -53,10 +40,6 @@ class VacancySource::Source::Fusion
   end
 
   private
-
-  def vacancy_listed_at_excluded_school_type?(schools)
-    (schools.map(&:detailed_school_type) & EXCLUDED_DETAILED_SCHOOL_TYPES).present?
-  end
 
   def attributes_for(item, schools)
     {
